Add 'configure-git' make target (#51)
This saves the user from copying and pasting a more complicated
one-liner from the README.

And also fix some missing .PHONY labels in Makefile, and normalize the
run_test.sh script -- no need to call go list ./..., and no need to
store the exit code and explicitly exit with it. Bash and sh exit with
the code of the file's last command.