Add config-server to kind
This is a lightweight service that will serve a config that has been
generated once. I preferred to generate the config once and mount the
volumes directly, but I was having trouble getting that to finish, so I
switched approaches to this.

Remaining work is to set up init containers on dheart, deyes, and sisud
to access http://config-server.config-server:8080/node$X/...?tar.gz=true
to download the appropriate set of files.

As of this commit, the sisu pods are still not configured so they do not
yet run.