Use kind for deployment (#58)
* Add create-kind-cluster make target

As of this commit, the cluster only runs two instances of ganache.

I use a Makefile for this because it is simple enough to orchestrate a
sequence of tasks; it is easy to use "make -j" to run many tasks in
parallel; and any workstation will have make available.

* Add mysql to kind cluster

I'm working on getting dheart running inside the cluster, but I have run
out of time tonight.

* Deploy deyes in kind

I don't think the absent chain configuration is correct, but it runs
this way for now.

* Temporarily use dheart build with bundled migrations

This is sisu-network/dheart#21, which I need for kind deployment at the
moment. Once the PR is merged, we can cut a release and depend on a
released version again.

* Deploy dheart in kind

* Deploy sisu to kind

It starts but crashes immediately due to invalid configuration.

And use a commit of dheart that is on the master branch, instead of the
temporary commit from a PR.

* Add kind/README.md

This attempts to address the first set of questions I anticipate a
contributor would have about using the kind cluster.

* Move ganache statefulset to its own namespace

To make this more clear that it is outside the scope of a Sisu instance.

* Remove namespace from kind .yml files

This requires the namespace to be specified explicitly when running
kubectl apply, which means it will be simple to deploy many full sisu
instances into the same kind cluster.

* Refactor kind/Makefile in preparation for multiple sisu instances

And rename the top-level make targets to kind-init-cluster and
kind-create-cluster, since we will have at least one new target in that
"namespace" soon.

* Add new make target "kind-add-sisu"

This adds a new instance of sisud+dheart+deyes to its own unique
namespace. At this point, sisud is still crashing due to lack of
configuration.

Also note that this moves the mysql instance to its own namespace. I
think the different dheart instances will need to use independent tables
before any real activity starts in this configuration.

* Add WIP init container for sisud instances in kind

The init container is intended to run "kubectl get ns" to discover the
other sisud instances in the kind cluster. I am currently getting an
RBAC error when that runs inside the cluster, even though
"auth can-i get ns --as=system:serviceaccount:sisu:namespace-lister"
reports that it is allowed.

* Fix clusterrolebinding for sisud in kind

Now the init container for sisud can successfully list other namespaces
in order to populate the sisud config.

* Create test keys and init sisud in kind

Now sisu instances in the kind cluster are started with:
sisud keys add --keyring-backend=test k1 && sisud init --chain-id=tbd sisu-1 --overwrite

which allows the sisud process to progress further, but now it fails
with:

Error: error during handshake: error on replay: validator set is nil in genesis and still empty after InitChain

I see some discussion on this error at
https://github.com/tendermint/tendermint/issues/5084, so I will have to
take a look later.

* Use buildkit cache mount to speed up kind docker builds

* Run single-instance sisu cluster inside kind

Using the new sisud testnet configuration.

Next, handle a dynamic count of sisud instances.

* Allow chain ID configuration from flag and set useOnMemory dependent on the number of peerId

* Deploy a configurable number of sisu instances to kind

They now use predictable namespaces. It defaults to 2 sisu instances,
but you can run 'make kind-init-cluster SISU_COUNT=4' for example, to
get 4 sisu instances in the kind cluster. Using the testnet config
creator, every sisu node should be aware of every other one.

* Add chain configs for the testnet

* Update sisud.yml

* Remove duplicated folder

* Use latest versions of dheart and deyes

* Add deyes-url entry to kind/include/sisu.toml

Without this, sisud was logging errors about:

cannot dial deyes, err = dial unix: missing address, sleeping before retry...

Now it gets further but fails with:

Error: error during handshake: error on replay: validator set is nil in genesis and still empty after InitChain

* Update kind/README.md to reflect current state

And add a section on how to get logs for a failing pod.

* Add script to to generate the data for all nodes in network

* Populate kind node with global sisu config

Next step is to copy the appropriate files or directories from the kind
node, into the pods. This will involve init containers similar to what
are present on sisud.

* Add dheart-config-volume and deyes-config-volume

* Revert "Add dheart-config-volume and deyes-config-volume"

This reverts commit 9aae1a21cfab20c9abbf5f435fe8a796bf9f901f.

I need to build off the previous commit.

* Add config-server to kind

This is a lightweight service that will serve a config that has been
generated once. I preferred to generate the config once and mount the
volumes directly, but I was having trouble getting that to finish, so I
switched approaches to this.

Remaining work is to set up init containers on dheart, deyes, and sisud
to access http://config-server.config-server:8080/node$X/...?tar.gz=true
to download the appropriate set of files.

As of this commit, the sisu pods are still not configured so they do not
yet run.

* Retrieve and mount configuration for sisu components

The Makefile doesn't have quite the right order of dependencies right
now, as timeouts are very likely. But even if make fails, you will
likely have a properly running cluster with this commit.

* Expand kind/Makefile dependencies

Don't attempt to deploy any sisu component until the config server has
been deployed, and don't deploy deyes until mysql has been deployed.

* Use correct ganache URLs in kind testnet

* Fix json tag name for deyes_ip

And bump deyes wait timeout to 60s, as 45s seems to be a boundary on
delay when restarting after an initial crash. The initial crash I've
seen a few times locally has been unable to resolve mysql.mysql, which I
think is a DNS propagation issue.

* Use zero-indexing for sisu namespaces in kind

This keeps it consistent with the testnet output of node0, node1, ...

* Expose p2p ports for dheart and sisud in kind

* Reorder steps in kind/Makefile

Deploy most services independently, but add separate "ready-" targets to
not block on the ready state of those services until the required
upstream services (MySQL, config server) are ready too.

* Expose ganache and sisud ports on localhost in kind

* Make shortcutPreparams in dheart param configuration when generating config files.

Co-authored-by: Benja Heart <visopsys@users.noreply.github.com>